requirements.txt
Flask==1.1.2
flask_bootstrap==3.3.7.1
flask_ckeditor==0.4.6
flask_gravatar==0.5.0
flask_login==0.6.2
flask_sqlalchemy==2.5.1
flask_wtf==1.0.1
python-dotenv==0.21.0
SQLAlchemy==1.4.22
Werkzeug==2.0.2
WTForms==3.0.1
